stable-requirements.txt
django-common
django-tagging
feedparser
simplejson
django-profiles
django-registration
paypal
django-pagination
  